Polish content for Powerpoint

Polish content for Powerpoint

# S3 method for class 'flextable'
polish_content_pptx(
  x,
  ph = "<p:ph/>",
  ...,
  full_width = FALSE,
  page_size = NULL,
  table_style = NULL,
  pptx = NULL,
  error_call = current_env()
)

# S3 method for class 'ggplot'
polish_content_pptx(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = 5,
  width = 6,
  units = "in",
  res = 300,
  error_call = current_env()
)

polish_content_pptx(x, ph = "<p:ph/>", pptx, ..., error_call = current_env())

# S3 method for class 'character'
polish_content_pptx(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  escape = TRUE,
  collapse = FALSE,
  font_color = NULL,
  font_style = NULL,
  font_size = NULL,
  font_typeface = NULL,
  error_call = current_env()
)

# S3 method for class 'list'
polish_content_pptx(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  list_type = c("none", "unordered", "ordered"),
  error_call = current_env()
)

# S3 method for class 'file_png'
polish_content_pptx(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = 5,
  width = 6,
  guess_size = FALSE,
  units = "in",
  error_call = current_env()
)

Arguments

x

object to be polished

ph

placeholder

...

These dots are for future extensions and must be empty.

full_width

Logical, used to fit full size of given width

page_size

Number of rows to accommodate in a slide

table_style

Named list with two entries: style and options. style is the name of table style to apply as a character, options a character vector of options to apply from the table style, including: "Banded Columns", "Banded Rows", "First Column", "Header Row", "Last Column", and "Total Row"

pptx

The powerpoint presentation.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

width, height, units

Plot size in units ("in", "cm", "mm", or "px"). defaults to 5 inches tall and 6 inches wide.

res

Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

escape

Should x be escaped with htmltools::htmlEscape()

collapse

If TRUE a <a:br w:type="textWrapping" w:clear="all"> element is added between the <a:r> nodes

font_color

color of text. Enter as a hex code.

font_style

styling of text, vector including "bold","italic","underline".

font_size

point size of text if different from document default.

font_typeface

typeface of the text to be added if different from document default.

list_type

How to print the list. Options are "none", "unordered" and "ordered".

"none" (default) puts contents on new lines and indents lists accordingly.

"unordered" makes the content bulleted list with indents.

"ordered" makes a numbered list with intents.

guess_size

see officer::external_img