Polish content for Word

Polish content for Word

# S3 method for class 'flextable'
polish_content_word(
  x,
  inline = TRUE,
  ...,
  align = NULL,
  split = NULL,
  topcaption = TRUE,
  keepnext = NULL,
  error_call = current_env()
)

# S3 method for class 'ggplot'
polish_content_word(
  x,
  inline = FALSE,
  ...,
  height = 5,
  width = 6,
  units = "in",
  res = 300,
  error_call = current_env()
)

polish_content_word(x, inline = FALSE, ..., error_call = current_env())

# S3 method for class 'character'
polish_content_word(
  x,
  inline = FALSE,
  ...,
  font_color = NULL,
  font_style = NULL,
  font_size = NULL,
  font_typeface = NULL,
  error_call = current_env()
)

# S3 method for class 'file_png'
polish_content_word(
  x,
  inline = FALSE,
  ...,
  height = 5,
  width = 6,
  units = "in",
  error_call = current_env()
)

Arguments

x

object to be polished

inline

if FALSE the comments are wrapped in a <w:p>

...

used in some methods

align

alignment of table and caption

split

should table be allowed to split

topcaption

should caption be above or below the table

keepnext

should content be marked for keep next.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

width, height, units

Plot size in units ("in", "cm", "mm", or "px"). defaults to 5 inches tall and 6 inches wide.

res

Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

font_color

color of text. Enter as a hex code.

font_style

styling of text, vector including "bold","italic","underline".

font_size

point size of text if different from document default.

font_typeface

typeface of the text to be added if different from document default.

Value

an object of class "polish_xml_nodeset". Contains XML that is viable Word OOXML for addition to Word documents.